distributing temoa_initialize into components
