+++ pyproject.toml
@@ -77,6 +77,9 @@ select = [
     "I",  # isort
     "B",  # flake8-bugbear
 ]
+ignore = [
+    "E501",  # line too long temporarily ignored
+]
 
 [tool.pytest.ini_options]
 minversion = "8.0"
