+++ temoa/_internal/temoa_initialize.py
@@ -19,12 +19,9 @@ in LICENSE.txt.  Users uncompressing this from an archive may not have
 received this license file.  If not, see <http://www.gnu.org/licenses/>.
 """
 
-from itertools import product as cross_product
-from operator import itemgetter as iget
 from sys import stderr as SE
-from typing import TYPE_CHECKING, Iterable
+from typing import TYPE_CHECKING
 
-from deprecated import deprecated
 from pyomo.core import Set
 
 if TYPE_CHECKING:
@@ -35,7 +32,7 @@ from logging import getLogger
 
 from pyomo.environ import value
 
-logger = getLogger(__name__)
+logger = getLogger(name=__name__)
 
 
 # ---------------------------------------------------------------
@@ -51,76 +48,7 @@ logger = getLogger(__name__)
 # ============================================================================
 __all__ = [
     # Index creation functions for constraints and variables
-    'AnnualCommodityBalanceConstraintIndices',
-    'AnnualRetirementVariableIndices',
-    'BaseloadDiurnalConstraintIndices',
-    'CapacityAnnualConstraintIndices',
-    'CapacityAvailableVariableIndices',
-    'CapacityConstraintIndices',
-    'CapacityFactorTechIndices',
-    'CapacityVariableIndices',
-    'CommodityBalanceConstraintIndices',
-    'CostFixedIndices',
-    'CostVariableIndices',
-    'CurtailmentVariableIndices',
-    'DemandActivityConstraintIndices',
-    'EmissionActivityIndices',
-    'FlexVariableAnnualIndices',
-    'FlexVariablelIndices',
-    'FlowInStorageVariableIndices',
-    'FlowVariableAnnualIndices',
-    'FlowVariableIndices',
-    'LifetimeLoanProcessIndices',
-    'LifetimeProcessIndices',
-    'LimitDegrowthCapacityIndices',
-    'LimitDegrowthNewCapacityDeltaIndices',
-    'LimitDegrowthNewCapacityIndices',
-    'LimitGrowthCapacityIndices',
-    'LimitGrowthNewCapacityDeltaIndices',
-    'LimitGrowthNewCapacityIndices',
-    'LimitTechInputSplitAnnualConstraintIndices',
-    'LimitTechInputSplitAverageConstraintIndices',
-    'LimitTechInputSplitConstraintIndices',
-    'LimitTechOutputSplitAnnualConstraintIndices',
-    'LimitTechOutputSplitAverageConstraintIndices',
-    'LimitTechOutputSplitConstraintIndices',
-    'LinkedTechConstraintIndices',
-    'ModelProcessLifeIndices',
-    'RampDownDayConstraintIndices',
-    'RampDownSeasonConstraintIndices',
-    'RampUpDayConstraintIndices',
-    'RampUpSeasonConstraintIndices',
-    'RegionalExchangeCapacityConstraintIndices',
-    'ReserveMarginIndices',
-    'RetiredCapacityVariableIndices',
-    'SeasonalStorageConstraintIndices',
-    'SeasonalStorageLevelVariableIndices',
-    'StorageConstraintIndices',
-    'StorageLevelVariableIndices',
-    # Validation and initialization functions
-    'CheckCapacityFactorProcess',
-    'CheckEfficiencyIndices',
-    'CheckEfficiencyVariable',
-    'CreateDemands',
-    'CreateRegionalIndices',
     'CreateSparseDicts',
-    'CreateSurvivalCurve',
-    'CreateTimeSeasonSequential',
-    'CreateTimeSequence',
-    # Default value functions
-    'get_default_capacity_factor',
-    'get_default_loan_rate',
-    'get_default_process_lifetime',
-    'get_default_survival',
-    'get_loan_life',
-    # Set initialization functions
-    'init_set_time_optimize',
-    'init_set_vintage_exist',
-    'init_set_vintage_optimize',
-    # Validation functions
-    'validate_SegFrac',
-    'validate_time',
-    'validate_TimeNext',
 ]
 
 
@@ -137,631 +65,6 @@ output carrier. Not currently used.
                 return True
 
     return False
-
-
-def get_str_padding(obj):
-    return len(str(obj))
-
-
-def CommodityBalanceConstraintErrorCheck(supplied, demanded, r, p, s, d, c):
-    # note:  if a pyomo equation simplifies to an int, there are no variables in it, which
-    #        is an indicator of a problem. How this might come up I do not know
-    if isinstance(supplied, int) or isinstance(demanded, int):
-        expr = str(supplied == demanded)
-        msg = (
-            'Unable to balance commodity {} in ({}, {}, {}, {}).\n'
-            'No flows on one side of constraint expression:\n'
-            '   {}\n'
-            'Possible reasons:\n'
-            " - Is there a missing period in set 'time_future'?\n"
-            " - Is there a missing tech in set 'tech_resource'?\n"
-            " - Is there a missing tech in set 'tech_production'?\n"
-            " - Is there a missing commodity in set 'commodity_physical'?\n"
-            ' - Are there missing entries in the Efficiency table?\n'
-            ' - Does a process need a longer Lifetime?'
-        )
-        logger.error(msg.format(c, r, p, s, d, expr))
-        raise Exception(msg.format(c, r, p, s, d, expr))
-
-
-def AnnualCommodityBalanceConstraintErrorCheck(supplied, demanded, r, p, c):
-    # note:  if a pyomo equation simplifies to an int, there are no variables in it, which
-    #        is an indicator of a problem. How this might come up I do not know
-    if isinstance(supplied, int) or isinstance(demanded, int):
-        expr = str(supplied == demanded)
-        msg = (
-            'Unable to balance annual commodity {} in ({}, {}).\n'
-            'No flows on one side of constraint expression:\n'
-            '   {}\n'
-            'Possible reasons:\n'
-            " - Is there a missing period in set 'time_future'?\n"
-            " - Is there a missing tech in set 'tech_resource'?\n"
-            " - Is there a missing tech in set 'tech_production'?\n"
-            " - Is there a missing commodity in set 'commodity_physical'?\n"
-            ' - Are there missing entries in the Efficiency table?\n'
-            ' - Does a process need a longer Lifetime?'
-        )
-        logger.error(msg.format(c, r, p, expr))
-        raise Exception(msg.format(c, r, p, expr))
-
-
-def DemandConstraintErrorCheck(supply, r, p, dem):
-    # note:  if a pyomo equation simplifies to an int, there are no variables in it, which
-    #        is an indicator of a problem
-    if isinstance(supply, int):
-        msg = (
-            "Error: Demand '{}' for ({}, {}) unable to be met by any "
-            'technology.\n\tPossible reasons:\n'
-            ' - Is the Efficiency parameter missing an entry for this demand?\n'
-            ' - Does a tech that satisfies this demand need a longer '
-            'Lifetime?\n'
-        )
-        logger.error(msg.format(dem, r, p))
-        raise Exception(msg.format(dem, r, p))
-
-
-def validate_time(M: 'TemoaModel'):
-    """
-    We check for integer status here, rather than asking Pyomo to do this via
-    a 'within=Integers' clause in the definition so that we can have a very
-    specific error message.  If we instead use Pyomo's mechanism, the
-    python invocation of Temoa throws an error (including a traceback)
-    that has proven to be scary and/or impenetrable for the typical modeler.
-    """
-    logger.debug('Started validating time index')
-    for year in M.time_exist:
-        if isinstance(year, int):
-            continue
-
-        msg = f'Set "time_exist" requires integer-only elements.\n\n  Invalid element: "{year}"'
-        logger.error(msg)
-        raise Exception(msg)
-
-    for year in M.time_future:
-        if isinstance(year, int):
-            continue
-
-        msg = f'Set "time_future" requires integer-only elements.\n\n invalid element: "{year}"'
-        logger.error(msg)
-        raise Exception(msg)
-
-    if len(M.time_future) < 2:
-        msg = (
-            'Set "time_future" needs at least 2 specified years.  \nTemoa '
-            'treats the integer numbers specified in this set as boundary years \n'
-            'between periods, and uses them to automatically ascertain the length \n'
-            '(in years) of each period.  Note that this means that there will be \n'
-            'one less optimization period than the number of elements in this set.'
-        )
-
-        logger.error(msg)
-        raise RuntimeError(msg)
-
-    # Ensure that the time_exist < time_future
-    if len(M.time_exist) > 0:
-        max_exist = max(M.time_exist)
-        min_horizon = min(M.time_future)
-
-        if not (max_exist < min_horizon):
-            msg = (
-                'All items in time_future must be larger than in time_exist.'
-                '\ntime_exist max:   {}'
-                '\ntime_future min: {}'
-            )
-            logger.error(msg.format(max_exist, min_horizon))
-            raise Exception(msg.format(max_exist, min_horizon))
-        logger.debug('Finished validating time')
-
-
-def validate_SegFrac(M: 'TemoaModel'):
-    """Ensure that the segment fractions adds up to 1"""
-
-    for p in M.time_optimize:
-        expected_keys = set((p, s, d) for s in M.TimeSeason[p] for d in M.time_of_day)
-        keys = set((_p, s, d) for _p, s, d in M.SegFrac.sparse_iterkeys() if _p == p)
-
-        if expected_keys != keys:
-            extra = keys.difference(expected_keys)
-            missing = expected_keys.difference(keys)
-            msg = (
-                'TimeSegmentFraction elements for period {} do not match TimeSeason and TimeOfDay.'
-                '\n\nIndices missing from TimeSegmentFraction:\n{}'
-                '\n\nIndices in TimeSegmentFraction missing from TimeSeason/TimeOfDay:\n{}'
-            ).format(p, missing, extra)
-            logger.error(msg)
-            raise ValueError(msg)
-
-        total = sum(M.SegFrac[k] for k in keys)
-
-        if abs(float(total) - 1.0) > 0.001:
-            # We can't explicitly test for "!= 1.0" because of incremental rounding
-            # errors associated with the specification of SegFrac by time slice,
-            # but we check to make sure it is within the specified tolerance.
-
-            key_padding = max(map(get_str_padding, keys))
-
-            fmt = '%%-%ds = %%s' % key_padding
-            # Works out to something like "%-25s = %s"
-
-            items = sorted([(k, M.SegFrac[k]) for k in keys])
-            items = '\n   '.join(fmt % (str(k), v) for k, v in items)
-
-            msg = (
-                'The values of TimeSegmentFraction do not sum to 1 for period {}. '
-                'Each item in SegFrac represents a fraction of a year, so they must '
-                'total to 1.  Current values:\n   {}\n\tsum = {}'
-            ).format(p, items, total)
-            logger.error(msg)
-            raise Exception(msg)
-
-
-def validate_TimeNext(M: 'TemoaModel'):
-    """
-    If using this table, check that defined states are actually valid.
-    TimeSegmentFraction is already compared to other tables so just compare to SegFrac.
-    """
-    # Only check TimeNext if it is actually being used
-    if M.TimeSequencing.first() != 'manual':
-        return
-
-    segfrac_psd = set(M.SegFrac.sparse_iterkeys())
-    time_next_psd = set((p, s, d) for p, s, d, s_next, d_next in M.TimeNext)
-    time_next_psd_next = set((p, s_next, d_next) for p, s, d, s_next, d_next in M.TimeNext)
-
-    missing_psd = segfrac_psd.difference(time_next_psd)
-    missing_psd_next = segfrac_psd.difference(time_next_psd_next)
-    if missing_psd or missing_psd_next:
-        msg = (
-            'Failed to build state sequence. '
-            '\nThese states from TimeSegmentFraction were not given a next state:\n{}\n'
-            '\nThese states from TimeSegmentFraction do not follow any state:\n{}'
-        ).format(missing_psd, missing_psd_next)
-        logger.error(msg)
-        raise ValueError(msg)
-
-
-def CheckEfficiencyIndices(M: 'TemoaModel'):
-    """
-    Ensure that there are no unused items in any of the Efficiency index sets.
-    """
-    # TODO:  This could be upgraded to scan for finer resolution
-    #        by checking by REGION and PERIOD...  Each region/period is unique.
-    c_physical = set(i for r, i, t, v, o in M.Efficiency.sparse_iterkeys())
-    c_physical = c_physical | set(i for r, i, t, v in M.ConstructionInput.sparse_iterkeys())
-    techs = set(t for r, i, t, v, o in M.Efficiency.sparse_iterkeys())
-    c_outputs = set(o for r, i, t, v, o in M.Efficiency.sparse_iterkeys())
-    c_outputs = c_outputs | set(o for r, t, v, o in M.EndOfLifeOutput.sparse_iterkeys())
-
-    symdiff = c_physical.symmetric_difference(M.commodity_physical)
-    if symdiff:
-        msg = (
-            'Unused or unspecified physical carriers.  Either add or remove '
-            'the following elements to the Set commodity_physical.'
-            '\n\n    Element(s): {}'
-        )
-        symdiff = (str(i) for i in symdiff)
-        f_msg = msg.format(', '.join(symdiff))
-        logger.error(f_msg)
-        raise ValueError(f_msg)
-
-    symdiff = techs.symmetric_difference(M.tech_all)
-    if symdiff:
-        msg = (
-            'Unused or unspecified technologies.  Either add or remove '
-            'the following technology(ies) to the tech_resource or '
-            'tech_production Sets.\n\n    Technology(ies): {}'
-        )
-        symdiff = (str(i) for i in symdiff)
-        f_msg = msg.format(', '.join(symdiff))
-        logger.error(f_msg)
-        raise ValueError(f_msg)
-
-    diff = M.commodity_demand - c_outputs
-    if diff:
-        msg = (
-            'Unused or unspecified outputs.  Either add or remove the '
-            'following elements to the commodity_demand Set.'
-            '\n\n    Element(s): {}'
-        )
-        diff = (str(i) for i in diff)
-        f_msg = msg.format(', '.join(diff))
-        logger.error(f_msg)
-        raise ValueError(f_msg)
-
-
-def CheckEfficiencyVariable(M: 'TemoaModel'):
-    count_rpitvo = dict()
-    # Pull non-variable efficiency by default
-    for r, i, t, v, o in M.Efficiency.sparse_iterkeys():
-        if (r, t, v) not in M.processPeriods:
-            # Probably an existing vintage that retires in p0
-            # Still want it for end of life flows
-            continue
-        for p in M.processPeriods[r, t, v]:
-            M.isEfficiencyVariable[r, p, i, t, v, o] = False
-            count_rpitvo[r, p, i, t, v, o] = 0
-
-    annual = set()
-    # Check for bad values and count up the good ones
-    for r, p, s, d, i, t, v, o in M.EfficiencyVariable.sparse_iterkeys():
-        if p not in M.processPeriods[r, t, v]:
-            msg = f'Invalid period {p} for process {r, t, v} in EfficiencyVariable table'
-            logger.error(msg)
-            raise ValueError(msg)
-
-        if t in M.tech_annual:
-            annual.add(t)
-
-        # Good value, pull from EfficiencyVariable table
-        count_rpitvo[r, p, i, t, v, o] += 1
-
-    for t in annual:
-        msg = (
-            f'Variable efficiencies were provided for the annual technology {t}, which has '
-            'no variable output. This will only be applied to flows on non-annual commodities. '
-            'This is ambiguous behaviour and not recommended.'
-        )
-        logger.warning(msg)
-
-    # Check if all possible values have been set as variable
-    # log a warning if some are missing (allowed but maybe accidental)
-    num_seg = len(M.TimeSeason[p]) * len(M.time_of_day)
-    for (r, p, i, t, v, o), count in count_rpitvo.items():
-        if count > 0:
-            M.isEfficiencyVariable[r, p, i, t, v, o] = True
-            if count < num_seg:
-                logger.info(
-                    'Some but not all EfficiencyVariable values were set (%i out of a possible %i) for: %s'
-                    ' Missing values will default to value set in Efficiency table.',
-                    count,
-                    num_seg,
-                    (r, p, i, t, v, o),
-                )
-
-
-def CheckCapacityFactorProcess(M: 'TemoaModel'):
-    count_rptv = dict()
-    # Pull CapacityFactorTech by default
-    for r, p, s, d, t in M.CapacityFactor_rpsdt:
-        for v in M.processVintages[r, p, t]:
-            M.isCapacityFactorProcess[r, p, t, v] = False
-            count_rptv[r, p, t, v] = 0
-
-    # Check for bad values and count up the good ones
-    for r, p, s, d, t, v in M.CapacityFactorProcess.sparse_iterkeys():
-        if v not in M.processVintages[r, p, t]:
-            msg = f'Invalid process {p, v} for {r, t} in CapacityFactorProcess table'
-            logger.error(msg)
-            raise ValueError(msg)
-
-        # Good value, pull from CapacityFactorProcess table
-        count_rptv[r, p, t, v] += 1
-
-    # Check if all possible values have been set by process
-    # log a warning if some are missing (allowed but maybe accidental)
-    for (r, p, t, v), count in count_rptv.items():
-        num_seg = len(M.TimeSeason[p]) * len(M.time_of_day)
-        if count > 0:
-            M.isCapacityFactorProcess[r, p, t, v] = True
-            if count < num_seg:
-                logger.info(
-                    'Some but not all processes were set in CapacityFactorProcess (%i out of a possible %i) for: %s'
-                    ' Missing values will default to CapacityFactorTech value or 1 if that is not set either.',
-                    count,
-                    num_seg,
-                    (r, p, t, v),
-                )
-
-
-@deprecated('should not be needed.  We are pulling the default on-the-fly where used')
-def CreateCapacityFactors(M: 'TemoaModel'):
-    """
-    Steps to creating capacity factors:
-    1. Collect all possible processes
-    2. Find the ones _not_ specified in CapacityFactorProcess
-    3. Set them, based on CapacityFactorTech.
-    """
-    # Shorter names, for us lazy programmer types
-    CFP = M.CapacityFactorProcess
-
-    # Step 1
-    processes = set((r, t, v) for r, i, t, v, o in M.Efficiency.sparse_iterkeys())
-
-    all_cfs = set(
-        (r, p, s, d, t, v)
-        for (r, t, v) in processes
-        for p in M.processPeriods[r, t, v]
-        for s, d in cross_product(M.TimeSeason[p], M.time_of_day)
-    )
-
-    # Step 2
-    unspecified_cfs = all_cfs.difference(CFP.sparse_iterkeys())
-
-    # Step 3
-
-    # Some hackery: We futz with _constructed because Pyomo thinks that this
-    # Param is already constructed.  However, in our view, it is not yet,
-    # because we're specifically targeting values that have not yet been
-    # constructed, that we know are valid, and that we will need.
-
-    if unspecified_cfs:
-        # CFP._constructed = False
-        for r, s, d, t, v in unspecified_cfs:
-            CFP[r, s, d, t, v] = M.CapacityFactorTech[r, s, d, t]
-        logger.debug(
-            'Created Capacity Factors for %d processes without an explicit specification',
-            len(unspecified_cfs),
-        )
-    # CFP._constructed = True
-
-
-def get_default_survival(M: 'TemoaModel', r, p, t, v):
-    """
-    Getting LifetimeSurvivalCurve where it is not defined
-    If this is a survival curve process, return 0 (likely beyond EOL)
-    Otherwise return 1 (no survival curve based EOL)
-    """
-    if M.isSurvivalCurveProcess[r, t, v]:
-        return 0
-    else:
-        return 1
-
-
-def get_default_process_lifetime(M: 'TemoaModel', r, t, v):
-    """
-    This initializer used to initialize the LifetimeProcess parameter from LifetimeTech where needed
-
-    Priority:
-        1.  Specified in LifetimeProcess data (provided as a fill and would not call this function)
-        2.  Specified in LifetimeTech data
-        3.  The default value from the LifetimeTech param (automatic)
-    :param M: generic model reference (not used)
-    :param r: region
-    :param t: tech
-    :param v: vintage
-    :return: the final lifetime value
-    """
-    return M.LifetimeTech[r, t]
-
-
-def get_default_capacity_factor(M: 'TemoaModel', r, p, s, d, t, v):
-    """
-    This initializer is used to fill the CapacityFactorProcess from the CapacityFactorTech where needed.
-
-    Priority:
-        1.  As specified in data input (this function not called)
-        2.  Here
-        3.  The default from CapacityFactorTech param
-    :param M: generic model reference
-    :param r: region
-    :param s: season
-    :param d: time-of-day slice
-    :param t: tech
-    :param v: vintage
-    :return: the capacity factor
-    """
-    return M.CapacityFactorTech[r, p, s, d, t]
-
-
-def get_default_loan_rate(M: 'TemoaModel', *_):
-    """get the default loan rate from the DefaultLoanRate param"""
-    return M.DefaultLoanRate()
-
-
-def CreateDemands(M: 'TemoaModel'):
-    """
-    Steps to create the demand distributions
-    1. Use Demand keys to ensure that all demands in commodity_demand are used
-    2. Find any slices not set in DemandDefaultDistribution, and set them based
-    on the associated SegFrac slice.
-    3. Validate that the DemandDefaultDistribution sums to 1.
-    4. Find any per-demand DemandSpecificDistribution values not set, and set
-    them from DemandDefaultDistribution.  Note that this only sets a
-    distribution for an end-use demand if the user has *not* specified _any_
-    anything for that end-use demand.  Thus, it is up to the user to fully
-    specify the distribution, or not.  No in-between.
-     5. Validate that the per-demand distributions sum to 1.
-    """
-    logger.debug('Started creating demand distributions in CreateDemands()')
-
-    # Step 0: some setup for a couple of reusable items
-    # Get the nth element from the tuple (r, p, s, d, dem)
-    # So we only have to update these indices in one place if they change
-    DSD_region = iget(0)
-    DSD_period = iget(1)
-    DSD_dem = iget(4)
-
-    # Step 1: Check if any demand commodities are going unused
-    used_dems = set(dem for r, p, dem in M.Demand.sparse_iterkeys())
-    unused_dems = sorted(M.commodity_demand.difference(used_dems))
-    if unused_dems:
-        for dem in unused_dems:
-            msg = "Warning: Demand '{}' is unused\n"
-            logger.warning(msg.format(dem))
-            SE.write(msg.format(dem))
-
-    # devnote: DDD just clones SegFrac. Unless we want to specify it in the database,
-    #          makes sense to just use SegFrac directly
-    # Step 2: Build the demand default distribution (= segfrac)
-    # DDD = M.DemandDefaultDistribution  # Shorter, for us lazy programmer types
-    # unset_defaults = set(M.SegFrac.sparse_iterkeys())
-    # unset_defaults.difference_update(DDD.sparse_iterkeys())
-    # if unset_defaults:
-    # Some hackery because Pyomo thinks that this Param is constructed.
-    # However, in our view, it is not yet, because we're specifically
-    # targeting values that have not yet been constructed, that we know are
-    # valid, and that we will need.
-    # DDD._constructed = False
-    # for tslice in unset_defaults:
-    #     DDD[tslice] = M.SegFrac[tslice]  # DDD._constructed = True
-
-    # Step 3: Check that DDD sums to 1
-    # devnote: this seems redundant to the SegFrac sum to 1 check.
-    # total = sum(i for i in DDD.values())
-    # if abs(value(total) - 1.0) > 0.001:
-    #     # We can't explicitly test for "!= 1.0" because of incremental rounding
-    #     # errors associated with the specification of demand shares by time slice,
-    #     # but we check to make sure it is within the specified tolerance.
-
-    #     key_padding = max(map(get_str_padding, DDD.sparse_iterkeys()))
-
-    #     fmt = '%%-%ds = %%s' % key_padding
-    #     # Works out to something like "%-25s = %s"
-
-    #     items = sorted(DDD.items())
-    #     items = '\n   '.join(fmt % (str(k), v) for k, v in items)
-
-    #     msg = (
-    #         'The values of the DemandDefaultDistribution parameter do not '
-    #         'sum to 1.  The DemandDefaultDistribution specifies how end-use '
-    #         'demands are distributed among the time slices (i.e., time_season, '
-    #         'time_of_day), so together, the data must total to 1.  Current '
-    #         'values:\n   {}\n\tsum = {}'
-    #     )
-    #     logger.error(msg.format(items, total))
-    #     raise ValueError(msg.format(items, total))
-
-    # Step 4: Fill out demand specific distribution table and check sums to 1 by region and demand
-    DSD = M.DemandSpecificDistribution
-
-    demands_specified = set(map(DSD_dem, (i for i in DSD.sparse_iterkeys())))
-    unset_demand_distributions = used_dems.difference(
-        demands_specified
-    )  # the demands not mentioned in DSD *at all*
-
-    if unset_demand_distributions:
-        for p in M.time_optimize:
-            unset_distributions = set(
-                cross_product(
-                    M.regions, (p,), M.TimeSeason[p], M.time_of_day, unset_demand_distributions
-                )
-            )
-            for r, p, s, d, dem in unset_distributions:
-                DSD[r, p, s, d, dem] = value(M.SegFrac[p, s, d])  # DSD._constructed = True
-
-    # Step 5: A final "sum to 1" check for all DSD members (which now should be everything)
-    #         Also check that all keys are made...  The demand distro should be supported
-    #         by the full set of (r, p, dem) keys because it is an equality constraint
-    #         and we need to ensure even the zeros are passed in
-    used_rp_dems = set((r, p, dem) for r, p, dem in M.Demand.sparse_iterkeys())
-    for r, p, dem in used_rp_dems:
-        expected_key_length = len(M.TimeSeason[p]) * len(M.time_of_day)
-        keys = [
-            k
-            for k in DSD.sparse_iterkeys()
-            if DSD_region(k) == r and DSD_period(k) == p and DSD_dem(k) == dem
-        ]
-        if len(keys) != expected_key_length:
-            # this could be very slow but only calls when there's a problem
-            missing = set(
-                (s, d)
-                for s in M.TimeSeason[p]
-                for d in M.time_of_day
-                if (r, p, s, d, dem) not in keys
-            )
-            logger.info(
-                'Missing some time slices for Demand Specific Distribution %s: %s',
-                (r, p, dem),
-                missing,
-            )
-        total = sum(value(DSD[i]) for i in keys)
-        if abs(value(total) - 1.0) > 0.001:
-            # We can't explicitly test for "!= 1.0" because of incremental rounding
-            # errors associated with the specification of demand shares by time slice,
-            # but we check to make sure it is within the specified tolerance.
-            key_padding = max(map(get_str_padding, keys))
-
-            fmt = '%%-%ds = %%s' % key_padding
-            # Works out to something like "%-25s = %s"
-
-            items = sorted((k, value(DSD[k])) for k in keys)
-            items = '\n   '.join(fmt % (str(k), v) for k, v in items)
-
-            msg = (
-                'The values of the DemandSpecificDistribution parameter do not '
-                'sum to 1 for {}. The DemandSpecificDistribution specifies how end-use '
-                'demands are distributed per time-slice (i.e., time_season, '
-                'time_of_day). Within each region, period, end-use demand, then, the distribution '
-                'must total to 1.\n\n Demand-specific distribution in error: '
-                ' \n   {}\n\tsum = {}'
-            )
-            logger.error(msg.format((r, p, dem), items, total))
-            raise ValueError(msg.format((r, p, dem), items, total))
-
-    logger.debug('Finished creating demand distributions')
-
-
-@deprecated(reason='vintage defaults are no longer available, so this should not be needed')
-def CreateCosts(M: 'TemoaModel'):
-    """
-    Steps to creating fixed and variable costs:
-    1. Collect all possible cost indices (CostFixed, CostVariable)
-    2. Find the ones _not_ specified in CostFixed and CostVariable
-    3. Set them, based on Cost*VintageDefault
-    """
-    logger.debug('Started Creating Fixed and Variable costs in CreateCosts()')
-    # Shorter names, for us lazy programmer types
-    CF = M.CostFixed
-    CV = M.CostVariable
-
-    # Step 1
-    fixed_indices = set(M.CostFixed_rptv)
-    var_indices = set(M.CostVariable_rptv)
-
-    # Step 2
-    unspecified_fixed_prices = fixed_indices.difference(CF.sparse_iterkeys())
-    unspecified_var_prices = var_indices.difference(CV.sparse_iterkeys())
-
-    # Step 3
-
-    # Some hackery: We futz with _constructed because Pyomo thinks that this
-    # Param is already constructed.  However, in our view, it is not yet,
-    # because we're specifically targeting values that have not yet been
-    # constructed, that we know are valid, and that we will need.
-
-    if unspecified_fixed_prices:
-        # CF._constructed = False
-        for r, p, t, v in unspecified_fixed_prices:
-            if (r, t, v) in M.CostFixedVintageDefault:
-                CF[r, p, t, v] = M.CostFixedVintageDefault[r, t, v]  # CF._constructed = True
-
-    if unspecified_var_prices:
-        # CV._constructed = False
-        for r, p, t, v in unspecified_var_prices:
-            if (r, t, v) in M.CostVariableVintageDefault:
-                CV[r, p, t, v] = M.CostVariableVintageDefault[r, t, v]
-    # CV._constructed = True
-    logger.debug('Created M.CostFixed with size: %d', len(M.CostFixed))
-    logger.debug('Created M.CostVariable with size: %d', len(M.CostVariable))
-    logger.debug('Finished creating Fixed and Variable costs')
-
-
-def init_set_time_optimize(M: 'TemoaModel'):
-    return sorted(M.time_future)[:-1]
-
-
-def init_set_vintage_exist(M: 'TemoaModel'):
-    return sorted(M.time_exist)
-
-
-def init_set_vintage_optimize(M: 'TemoaModel'):
-    return sorted(M.time_optimize)
-
-
-def CreateRegionalIndices(M: 'TemoaModel'):
-    """Create the set of all regions and all region-region pairs"""
-    regional_indices = set()
-    for r_i in M.regions:
-        if '-' in r_i:
-            logger.error("Individual region names can not have '-' in their names: %s", str(r_i))
-            raise ValueError("Individual region names can not have '-' in their names: " + str(r_i))
-        for r_j in M.regions:
-            if r_i == r_j:
-                regional_indices.add(r_i)
-            else:
-                regional_indices.add(r_i + '-' + r_j)
-    # dev note:  Sorting these passed them to pyomo in an ordered container and prevents warnings
-    return sorted(regional_indices)
 
 
 # ---------------------------------------------------------------
@@ -1170,342 +473,11 @@ def CreateSparseDicts(M: 'TemoaModel'):
     logger.debug('Completed creation of SparseDicts')
 
 
-def CreateTimeSequence(M: 'TemoaModel'):
-    logger.debug('Creating sequence of time slices.')
-
-    # Establishing sequence of states
-    match M.TimeSequencing.first():
-        case 'consecutive_days':
-            msg = 'Running a consecutive days database.'
-            for p in M.time_optimize:
-                for s, d in M.TimeSeason[p] * M.time_of_day:
-                    M.time_next[p, s, d] = loop_period_next_timeslice(M, p, s, d)
-        case 'seasonal_timeslices':
-            msg = 'Running a seasonal time slice database.'
-            for p in M.time_optimize:
-                for s, d in M.TimeSeason[p] * M.time_of_day:
-                    M.time_next[p, s, d] = loop_season_next_timeslice(M, p, s, d)
-        case 'representative_periods':
-            msg = 'Running a representative periods database.'
-            for p in M.time_optimize:
-                for s, d in M.TimeSeason[p] * M.time_of_day:
-                    M.time_next[p, s, d] = loop_season_next_timeslice(M, p, s, d)
-        case 'manual':
-            # Hidden feature. Define the sequence directly in the TimeNext table
-            msg = 'Pulling time sequence from TimeNext table.'
-            for p, s, d, s_next, d_next in M.TimeNext:
-                M.time_next[p, s, d] = s_next, d_next
-        case _:
-            # This should have been caught in hybrid_loader
-            msg = f"Invalid time sequencing parameter loaded '{M.TimeSequencing.first()}'. Likely code error."
-            logger.error(msg)
-            raise ValueError(msg)
-
-    msg += ' This behaviour can be changed using the time_sequencing parameter in the config file. '
-    logger.info(msg)
-
-    logger.debug('Creating superimposed sequential seasons.')
-
-    # Superimposed sequential seasons
-    for p in M.time_optimize:
-        seasons = [(s_seq, s) for _p, s_seq, s in M.ordered_season_sequential if _p == p]
-        for i, (s_seq, s) in enumerate(seasons):
-            M.sequential_to_season[p, s_seq] = s
-            if (s_seq, s) == seasons[-1]:
-                M.time_next_sequential[p, s_seq] = seasons[0][0]
-            else:
-                M.time_next_sequential[p, s_seq] = seasons[i + 1][0]
-
-    logger.debug('Created time sequence.')
-
-
-def CreateTimeSeasonSequential(M: 'TemoaModel'):
-    if all(
-        (
-            not M.tech_seasonal_storage,
-            not M.RampUpHourly,
-            not M.RampDownHourly,
-        )
-    ):
-        # Don't need it anyway
-        return
-
-    if not M.TimeSeasonSequential:
-        if M.TimeSequencing.first() in ('consecutive_days', 'seasonal_timeslices'):
-            logger.info(
-                'No data in TimeSeasonSequential. By default, assuming sequential seasons '
-                'match TimeSeason and TimeSegmentFraction.'
-            )
-            for s in M.time_season:
-                M.time_season_sequential.add(s)
-            for p in M.TimeSeason:
-                for s in M.TimeSeason[p]:
-                    M.ordered_season_sequential.add((p, s, s))
-                    M.TimeSeasonSequential[p, s, s] = value(M.SegFracPerSeason[p, s]) * value(
-                        M.DaysPerPeriod
-                    )
-
-        else:
-            msg = (
-                f'No data in TimeSeasonSequential but time_sequencing parameter set to {M.TimeSequencing.first()} '
-                'and inter-season features used. TimeSeasonSequential must be filled for this type of time '
-                'sequencing if seasonal storage or inter-season constraints like RampUp/RampDown are used. Check '
-                'the config file.'
-            )
-            logger.error(msg)
-            raise ValueError(msg)
-
-    sequential = dict()
-    prev_n = 0
-    for p, s_seq, s in M.TimeSeasonSequential.sparse_iterkeys():
-        num_days = value(M.TimeSeasonSequential[p, s_seq, s])
-        if (
-            M.TimeSequencing.first() == 'consecutive_days'
-            and prev_n
-            and abs(num_days - prev_n) >= 0.001
-        ):
-            msg = (
-                'TimeSequencing set to consecutive_days but two consecutive seasons do not represent the same '
-                f'number of days. This discontinuity will lead to bad model behaviour: {p, s}, days: {num_days}. '
-                f'Previous number of days: {prev_n}. Check the config file for more information.'
-            )
-            logger.error(msg)
-            raise ValueError(msg)
-        prev_n = num_days  # for validating next in sequence
-
-        # Regardless of their order, make sure the total number of days adds up
-        if (p, s) not in sequential:
-            sequential[p, s] = 0
-        sequential[p, s] += num_days
-
-    # Check that TimeSeasonSequential num_days total to number of days in each period
-    count_total = dict()  # {p: n} total days per period according to TimeSeasonSequential
-    for p in M.time_optimize:
-        count_total[p] = sum(sequential[p, s] for _p, s in sequential if _p == p)
-        if abs(count_total[p] - value(M.DaysPerPeriod)) >= 0.001:
-            logger.warning(
-                f'Sum of num_days in TimeSeasonSequential ({count_total[p]}) '
-                f'for period {p} does not sum to days_per_period ({value(M.DaysPerPeriod)}) '
-                'from the MetaData table.'
-            )
-
-    # Check that seasons using in storage seasons are actual seasons
-    for p, s in sequential:
-        if (p, s) not in M.SegFracPerSeason:
-            msg = (
-                f'Period-season index {(p, s)} that does not exist in '
-                'TimeSegmentFraction referenced in TimeSeasonSequential .'
-            )
-            logger.error(msg)
-            raise ValueError(msg)
-
-    for p, s in M.SegFracPerSeason.sparse_iterkeys():
-        if s not in M.TimeSeason[p]:
-            continue
-
-        # Check that all seasons are used in sequential seasons
-        if (p, s) not in sequential:
-            msg = f'Period-season index {(p, s)} absent from TimeSeasonSequential'
-            logger.warning(msg)
-
-        # Check that the two tables agree on the total seasonal composition of each period
-        segfrac = value(M.SegFracPerSeason[p, s])
-        segfracseq = sequential[p, s] / count_total[p]
-        if abs(segfrac - segfracseq) >= 0.001:
-            msg = (
-                'Discrepancy of total period-season composition between '
-                'TimeSegmentFraction and TimeSeasonSequential. Total fraction of each '
-                'period assigned to each season should match: '
-                f'TimeSegmentFraction: {(p, s, value(M.SegFracPerSeason[p, s]))}'
-                f', TimeSeasonSequential: {(p, s, segfracseq)}'
-            )
-            logger.warning(msg)
-
-
-def CreateSurvivalCurve(M: 'TemoaModel'):
-    rtv_interpolated = set()  # so we only need one warning
-
-    for r, _, t, v, _ in M.Efficiency.sparse_iterkeys():
-        M.isSurvivalCurveProcess[r, t, v] = False  # by default
-
-    # Collect rptv indices into (r, t, v): p dictionary
-    for r, p, t, v in M.LifetimeSurvivalCurve.sparse_iterkeys():
-        if (r, t, v) not in M.survivalCurvePeriods:
-            M.survivalCurvePeriods[r, t, v] = list()
-        M.survivalCurvePeriods[r, t, v].append(p)
-        M.isSurvivalCurveProcess[r, t, v] = True
-
-    # Go through all the periods for each (r, t, v) in order
-    for r, t, v in M.survivalCurvePeriods:
-        periods_rtv = sorted(M.survivalCurvePeriods[r, t, v])
-
-        p_first = periods_rtv[0]
-        p_last = periods_rtv[-1]
-
-        if p_first != v:
-            msg = (
-                'LifetimeSurvivalCurve must be defined starting in the vintage period. Must '
-                f'define ({r}, >{v}<, {t}, {v})'
-            )
-            logger.error(msg)
-            raise ValueError(msg)
-
-        if value(M.LifetimeSurvivalCurve[r, v, t, v]) != 1:
-            msg = (
-                'LifetimeSurvivalCurve must begin at 1 for calculating annual retirements. ',
-                f'Got {value(M.LifetimeSurvivalCurve[r, v, t, v])} for ({r}, {v}, {t}, {v})',
-            )
-            logger.error(msg)
-            raise ValueError(msg)
-
-        # Collect a list of processes that needed to be interpolated, for warning
-        if periods_rtv != list(range(p_first, p_last + 1, 1)):
-            rtv_interpolated.add((r, t, v))
-
-        between_periods = []
-        for i, p in enumerate(periods_rtv):
-            if i == 0:
-                continue  # Cant look back from first period. Could be zero but hey why not
-
-            # Check that the survival curve monotonically decreases
-            p_prev = periods_rtv[i - 1]
-            lsc = value(M.LifetimeSurvivalCurve[r, p, t, v])
-            lsc_prev = value(M.LifetimeSurvivalCurve[r, p_prev, t, v])
-            if lsc - lsc_prev > 0.0001:
-                msg = (
-                    'LifetimeSurvivalCurve fraction increases going forward in time from {} to {}. '
-                    'This is not allowed.'
-                ).format((r, p_prev, t, v), (r, p, t, v))
-                logger.error(msg)
-                raise ValueError(msg)
-
-            if p - p_prev > 1:
-                _between_periods = list(range(p_prev + 1, p, 1))
-                for _p in _between_periods:
-                    x = (_p - p_prev) / (p - p_prev)
-                    lsc_x = lsc_prev + x * (lsc - lsc_prev)
-                    M.LifetimeSurvivalCurve[r, _p, t, v] = lsc_x
-                between_periods.extend(_between_periods)
-
-            if lsc < 0.0001:
-                if p != p_last:
-                    msg = (
-                        'There is no need to continue a survival curve beyond fraction ~= 0. '
-                        f'ignoring periods beyond {p} for ({r, t, v})'
-                    )
-                    logger.info(msg)
-
-                # Make sure the lifetime for this process aligns with survival curve end
-                if value(M.LifetimeProcess[r, t, v]) < p - v:
-                    msg = (
-                        f'The LifetimeProcess parameter for process ({r, t, v}) with survival curve  '
-                        f'does not extend beyond the end of that survival curve in {p}. To agree with '
-                        f'the survival curve, set LifetimeProcess[{r, t, v}] >= {p - v}'
-                    )
-                    logger.error(msg)
-                    raise ValueError(msg)
-                elif value(M.LifetimeProcess[r, t, v]) != p - v:
-                    msg = (
-                        f'The LifetimeProcess parameter for process ({r, t, v}) with survival curve  '
-                        f'does match the end of that survival curve in {p}. This will waste compute. '
-                        'To agree with the survival curve and suppress this warning, set '
-                        f'LifetimeProcess[{r, t, v}] = {p - v}'
-                    )
-                    logger.warning(msg)
-
-                continue
-
-            # Flag if the last period is not fraction = 0. This is important for investment costs
-            if p == p_last and lsc > 0.0001:
-                msg = (
-                    'Any defined survival curve must continue to zero for the purposes of '
-                    'investment cost accounting, even if this period would extend beyond '
-                    f'defined future periods. Continue ({r, t, v}) to fraction == 0.'
-                )
-                logger.error(msg)
-                raise ValueError(msg)
-
-        M.survivalCurvePeriods[r, t, v].extend(between_periods)
-        M.survivalCurvePeriods[r, t, v] = set(M.survivalCurvePeriods[r, t, v])
-
-    if rtv_interpolated:
-        msg = (
-            'For the purposes of investment cost accounting, LifetimeSurvivalCurve must be defined '
-            'for each individual year. Gaps between defined years will be filled by linear interpolation. '
-            'Otherwise, these individual years can be defined manually. Interpolated processes: {}'
-        ).format([rtv for rtv in rtv_interpolated])
-        logger.info(msg)
-
-
 # ---------------------------------------------------------------
 # Create sparse parameter indices.
 # These functions are called from temoa_model.py and use the sparse keys
 # associated with specific parameters.
 # ---------------------------------------------------------------
-
-
-@deprecated('switched over to validator... this set is typically VERY empty')
-def CapacityFactorProcessIndices(M: 'TemoaModel'):
-    indices = set(
-        (r, s, d, t, v)
-        for r, i, t, v, o in M.Efficiency.sparse_iterkeys()
-        for p in M.time_optimize
-        for s in M.TimeSeason[p]
-        for d in M.time_of_day
-    )
-    return indices
-
-
-def CapacityFactorTechIndices(M: 'TemoaModel'):
-    all_cfs = set(
-        (r, p, s, d, t)
-        for r, p, t in M.activeCapacityAvailable_rpt
-        for s in M.TimeSeason[p]
-        for d in M.time_of_day
-    )
-    return all_cfs
-
-
-def CostFixedIndices(M: 'TemoaModel'):
-    # we pull the unlimited capacity techs from this index.  They cannot have fixed costs
-    return {(r, p, t, v) for r, p, t, v in M.activeActivity_rptv if t not in M.tech_uncap}
-
-
-def CostVariableIndices(M: 'TemoaModel'):
-    return M.activeActivity_rptv
-
-
-# dev note:  appears superfluous...
-# def CostInvestIndices(M: 'TemoaModel'):
-#     indices = set((r, t, v) for r, p, t, v in M.processLoans)
-#
-#     return indices
-
-
-@deprecated('No longer used.  See the region_group_check in validators.py')
-def RegionalGlobalInitializedIndices(M: 'TemoaModel'):
-    from itertools import permutations
-
-    indices = set()
-    for n in range(1, len(M.regions) + 1):
-        regional_perms = permutations(M.regions, n)
-        for i in regional_perms:
-            indices.add('+'.join(i))
-    indices.add('global')
-    indices = indices.union(M.regionalIndices)
-
-    return indices
-
-
-def EmissionActivityIndices(M: 'TemoaModel'):
-    indices = set(
-        (r, e, i, t, v, o)
-        for r, i, t, v, o in M.Efficiency.sparse_iterkeys()
-        for e in M.commodity_emissions
-        if r in M.regions  # omit any exchange/groups
-    )
-
-    return indices
 
 
 # devnote: this does not appear to be used anywhere
@@ -1518,589 +490,11 @@ def EmissionActivityIndices(M: 'TemoaModel'):
 #     return indices
 
 
-def ModelProcessLifeIndices(M: 'TemoaModel'):
-    """\
-Returns the set of sensical (region, period, tech, vintage) tuples.  The tuple indicates
-the periods in which a process is active, distinct from TechLifeFracIndices that
-returns indices only for processes that EOL mid-period.
-"""
-    return M.activeActivity_rptv
-
-
-def LifetimeProcessIndices(M: 'TemoaModel'):
-    """\
-Based on the Efficiency parameter's indices, this function returns the set of
-process indices that may be specified in the LifetimeProcess parameter.
-"""
-    indices = set((r, t, v) for r, i, t, v, o in M.Efficiency.sparse_iterkeys())
-
-    return indices
-
-
-def LifetimeLoanProcessIndices(M: 'TemoaModel'):
-    """\
-Based on the Efficiency parameter's indices and time_future parameter, this
-function returns the set of process indices that may be specified in the
-CostInvest parameter.
-"""
-    min_period = min(M.vintage_optimize)
-
-    indices = set((r, t, v) for r, i, t, v, o in M.Efficiency.sparse_iterkeys() if v >= min_period)
-
-    return indices
-
-
 # ---------------------------------------------------------------
 # Create sparse indices for decision variables.
 # These functions are called from temoa_model.py and use the dictionaries
 # created above in CreateSparseDicts()
 # ---------------------------------------------------------------
-
-
-def CapacityVariableIndices(M: 'TemoaModel'):
-    return M.newCapacity_rtv
-
-
-def RetiredCapacityVariableIndices(M: 'TemoaModel'):
-    return set(
-        (r, p, t, v)
-        for r, p, t in M.processVintages
-        if t in M.tech_retirement and t not in M.tech_uncap
-        for v in M.processVintages[r, p, t]
-        if v < p <= v + value(M.LifetimeProcess[r, t, v]) - value(M.PeriodLength[p])
-    )
-
-
-def AnnualRetirementVariableIndices(M: 'TemoaModel'):
-    return set(
-        (r, p, t, v) for r, t, v in M.retirementPeriods for p in M.retirementPeriods[r, t, v]
-    )
-
-
-def CapacityAvailableVariableIndices(M: 'TemoaModel'):
-    return M.activeCapacityAvailable_rpt
-
-
-def CapacityAvailableVariableIndicesVintage(M: 'TemoaModel'):
-    return M.activeCapacityAvailable_rptv
-
-
-def FlowVariableIndices(M: 'TemoaModel'):
-    return M.activeFlow_rpsditvo
-
-
-def FlowVariableAnnualIndices(M: 'TemoaModel'):
-    return M.activeFlow_rpitvo
-
-
-def FlexVariablelIndices(M: 'TemoaModel'):
-    return M.activeFlex_rpsditvo
-
-
-def FlexVariableAnnualIndices(M: 'TemoaModel'):
-    return M.activeFlex_rpitvo
-
-
-def FlowInStorageVariableIndices(M: 'TemoaModel'):
-    return M.activeFlowInStorage_rpsditvo
-
-
-def CurtailmentVariableIndices(M: 'TemoaModel'):
-    return M.activeCurtailment_rpsditvo
-
-
-def StorageLevelVariableIndices(M: 'TemoaModel'):
-    return M.storageLevelIndices_rpsdtv
-
-
-def SeasonalStorageLevelVariableIndices(M: 'TemoaModel'):
-    return M.seasonalStorageLevelIndices_rpstv
-
-
-def SeasonalStorageConstraintIndices(M: 'TemoaModel'):
-    indices = set(
-        (r, p, s, d, t, v)
-        for r, p, s, t, v in M.seasonalStorageLevelIndices_rpstv
-        for d in M.time_of_day
-    )
-
-    return indices
-
-
-def StorageConstraintIndices(M: 'TemoaModel'):
-    return M.storageLevelIndices_rpsdtv
-
-
-def CapacityConstraintIndices(M: 'TemoaModel'):
-    capacity_indices = set(
-        (r, p, s, d, t, v)
-        for r, p, t, v in M.activeActivity_rptv
-        if (t not in M.tech_annual or t in M.tech_demand)
-        if t not in M.tech_uncap
-        if t not in M.tech_storage
-        for s in M.TimeSeason[p]
-        for d in M.time_of_day
-    )
-
-    return capacity_indices
-
-
-def LinkedTechConstraintIndices(M: 'TemoaModel'):
-    linkedtech_indices = set(
-        (r, p, s, d, t, v, e)
-        for r, t, e in M.LinkedTechs.sparse_iterkeys()
-        for p in M.time_optimize
-        if (r, p, t) in M.processVintages
-        for v in M.processVintages[r, p, t]
-        if (r, p, t, v) in M.activeActivity_rptv
-        for s in M.TimeSeason[p]
-        for d in M.time_of_day
-    )
-
-    return linkedtech_indices
-
-
-def CapacityAnnualConstraintIndices(M: 'TemoaModel'):
-    capacity_indices = set(
-        (r, p, t, v)
-        for r, p, t, v in M.activeActivity_rptv
-        if t in M.tech_annual and t not in M.tech_demand
-        if t not in M.tech_uncap
-    )
-
-    return capacity_indices
-
-
-# ---------------------------------------------------------------
-# Create sparse indices for constraints.
-# These functions are called from temoa_model.py and use the dictionaries
-# created above in CreateSparseDicts()
-# ---------------------------------------------------------------
-
-
-# def DemandActivityConstraintIndices(M: 'TemoaModel'):
-#     """\
-# This function returns a set of sparse indices that are used in the
-# DemandActivity constraint. It returns a tuple of the form:
-# (p,s,d,t,v,dem,first_s,first_d) where "dem" is a demand commodity, and "first_s"
-# and "first_d" are the reference season and time-of-day, respectively used to
-# ensure demand activity remains consistent across time slices.
-# """
-
-#     # needed data structures...
-#     # the count of techs that supply a commodity
-#     suppliers = defaultdict(set)
-#     # (region, demand): (season, tod)  # the goal of the exercise!
-#     anchor_season_tod = {}
-#     # (region, demand): (period, tech, vintage) # the viable tech and vintage per region, demand
-#     viable_tech_vintage = defaultdict(list)
-
-#     # start the loop over possible combos
-#     for r, p, t, v, dem in M.processInputsByOutput:
-#         # we aren't concerned with non-demand commodities or annual techs
-#         if dem not in M.commodity_demand or t in M.tech_annual:
-#             continue
-#         # capture the (p, t, v) in case we need to act on it
-#         viable_tech_vintage[r, p, dem].append((t, v))
-#         suppliers[dem].add(t)  # one more recognized supplier
-#         if len(suppliers[dem]) > 1:
-#             # We need to act on (build) for this region-demand, put in a placeholder
-#             anchor_season_tod[r, p, dem] = None
-
-#     # Find the first timestep of the year where the demand is appreciably sized:
-#     #   appreciable = not so small that we get into numerical instability when applying small multipliers
-#     appreciable_size = 0.0001
-
-#     for r, p, dem in anchor_season_tod:
-#         found_flag = False
-#         s0, d0 = None, None
-#         for s0, d0 in ((ss, dd) for ss in M.TimeSeason[p] for dd in M.time_of_day):
-#             if (r, p, s0, d0, dem) in M.DemandSpecificDistribution:
-#                 if value(M.DemandSpecificDistribution[r, p, s0, d0, dem]) >= appreciable_size:
-#                     found_flag = True
-#                     break  # we have one with some value associated
-#         found = 'found' if found_flag else 'not found'
-#         # set it.  If nothing was found the first indices should work just fine...
-#         anchor_season_tod[r, p, dem] = (s0, d0)
-#         logger.debug(
-#             'Using season/tod: %s, %s for commodity %s in region %s which was %s in DSD '
-#             'to set DemandActivity baseline',
-#             s0,
-#             d0,
-#             dem,
-#             r,
-#             found,
-#         )
-
-#     # Start yielding the constraint indices
-#     for r, p, dem in anchor_season_tod:
-#         s0, d0 = anchor_season_tod[r, p, dem]
-#         for t, v in viable_tech_vintage[r, p, dem]:
-#             for s in M.TimeSeason[p]:
-#                 for d in M.time_of_day:
-#                     if s != s0 or d != d0:
-#                         yield r, p, s, d, t, v, dem, s0, d0
-
-
-def DemandActivityConstraintIndices(M: 'TemoaModel'):
-    indices = set(
-        (r, p, s, d, t, v, dem)
-        for r, p, dem in M.DemandConstraint_rpc
-        for t, v in M.commodityUStreamProcess[r, p, dem]
-        if t not in M.tech_annual
-        for s in M.TimeSeason[p]
-        for d in M.time_of_day
-    )
-    return indices
-
-
-# devnote: no longer needed
-# def DemandConstraintIndices(M: 'TemoaModel'):
-#     indices = set(
-#         (r, p, dem)
-#         for r, p, dem in M.Demand.sparse_iterkeys()
-#     )
-
-#     return indices
-
-
-def BaseloadDiurnalConstraintIndices(M: 'TemoaModel'):
-    indices = set(
-        (r, p, s, d, t, v)
-        for r, p, t in M.baseloadVintages
-        for v in M.baseloadVintages[r, p, t]
-        for s in M.TimeSeason[p]
-        for d in M.time_of_day
-    )
-
-    return indices
-
-
-def RegionalExchangeCapacityConstraintIndices(M: 'TemoaModel'):
-    indices = set(
-        (r_e, r_i, p, t, v)
-        for r_e, p, i in M.exportRegions
-        for r_i, t, v, o in M.exportRegions[r_e, p, i]
-    )
-
-    return indices
-
-
-def CommodityBalanceConstraintIndices(M: 'TemoaModel'):
-    # Generate indices only for those commodities that are produced by
-    # technologies with varying output at the time slice level.
-    indices = set(
-        (r, p, s, d, c)
-        for r, p, c in M.commodityBalance_rpc
-        # r in this line includes interregional transfer combinations (not needed).
-        if r in M.regions  # this line ensures only the regions are included.
-        and c not in M.commodity_annual
-        for s in M.TimeSeason[p]
-        for d in M.time_of_day
-    )
-
-    return indices
-
-
-def AnnualCommodityBalanceConstraintIndices(M: 'TemoaModel'):
-    # Generate indices only for those commodities that are produced by
-    # technologies with constant annual output.
-    indices = set(
-        (r, p, c)
-        for r, p, c in M.commodityBalance_rpc
-        # r in this line includes interregional transfer combinations (not needed).
-        if r in M.regions  # this line ensures only the regions are included.
-        and c in M.commodity_annual
-    )
-
-    return indices
-
-
-def RampUpDayConstraintIndices(M: 'TemoaModel'):
-    indices = set(
-        (r, p, s, d, t, v)
-        for r, p, t in M.rampUpVintages
-        for v in M.rampUpVintages[r, p, t]
-        for s in M.TimeSeason[p]
-        for d in M.time_of_day
-    )
-
-    return indices
-
-
-def RampDownDayConstraintIndices(M: 'TemoaModel'):
-    indices = set(
-        (r, p, s, d, t, v)
-        for r, p, t in M.rampDownVintages
-        for v in M.rampDownVintages[r, p, t]
-        for s in M.TimeSeason[p]
-        for d in M.time_of_day
-    )
-
-    return indices
-
-
-def RampUpSeasonConstraintIndices(M: 'TemoaModel'):
-    if M.TimeSequencing.first() == 'consecutive_days':
-        return Set.Skip  # dont need this constraint
-
-    # s, s_next indexing ensures we dont build redundant constraints
-    indices = set(
-        (r, p, s, s_next, t, v)
-        for r, p, t in M.rampUpVintages
-        for v in M.rampUpVintages[r, p, t]
-        for _p, s_seq, s in M.ordered_season_sequential
-        if _p == p
-        for s_seq_next in (M.time_next_sequential[p, s_seq],)  # next sequential season
-        for s_next in (
-            M.sequential_to_season[p, s_seq_next],
-        )  # next sequential season's matching season
-        if s_next
-        != M.time_next[p, s, M.time_of_day.last()][0]  # to avoid redundancy on RampDay constraint
-    )
-
-    return indices
-
-
-def RampDownSeasonConstraintIndices(M: 'TemoaModel'):
-    if M.TimeSequencing.first() == 'consecutive_days':
-        return Set.Skip  # dont need this constraint
-
-    # s, s_next indexing ensures we dont build redundant constraints
-    indices = set(
-        (r, p, s, s_next, t, v)
-        for r, p, t in M.rampDownVintages
-        for v in M.rampDownVintages[r, p, t]
-        for _p, s_seq, s in M.ordered_season_sequential
-        for s_seq_next in (M.time_next_sequential[p, s_seq],)  # next sequential season
-        for s_next in (
-            M.sequential_to_season[p, s_seq_next],
-        )  # next sequential season's matching season
-        if s_next
-        != M.time_next[p, s, M.time_of_day.last()][0]  # to avoid redundancy on RampDay constraint
-    )
-
-    return indices
-
-
-def ReserveMarginIndices(M: 'TemoaModel'):
-    indices = set(
-        (r, p, s, d)
-        for r in M.PlanningReserveMargin.sparse_iterkeys()
-        for p in M.time_optimize
-        if (r, p) in M.processReservePeriods
-        for s in M.TimeSeason[p]
-        for d in M.time_of_day
-    )
-
-    return indices
-
-
-def LimitTechInputSplitConstraintIndices(M: 'TemoaModel'):
-    indices = set(
-        (r, p, s, d, i, t, v, op)
-        for r, p, i, t, op in M.inputSplitVintages
-        if t not in M.tech_annual
-        for v in M.inputSplitVintages[r, p, i, t, op]
-        for s in M.TimeSeason[p]
-        for d in M.time_of_day
-    )
-    ann_indices = set(
-        (r, p, i, t, op) for r, p, i, t, op in M.inputSplitVintages if t in M.tech_annual
-    )
-    if len(ann_indices) > 0:
-        msg = (
-            'Warning: Annual technologies included in LimitTechInputSplit table. '
-            'Use LimitTechInputSplitAnnual table instead or these constraints will be ignored: {}'
-        )
-        logger.warning(msg.format(ann_indices))
-
-    return indices
-
-
-def LimitTechInputSplitAnnualConstraintIndices(M: 'TemoaModel'):
-    indices = set(
-        (r, p, i, t, v, op)
-        for r, p, i, t, op in M.inputSplitAnnualVintages
-        if t in M.tech_annual
-        for v in M.inputSplitAnnualVintages[r, p, i, t, op]
-    )
-
-    return indices
-
-
-def LimitTechInputSplitAverageConstraintIndices(M: 'TemoaModel'):
-    indices = set(
-        (r, p, i, t, v, op)
-        for r, p, i, t, op in M.inputSplitAnnualVintages
-        if t not in M.tech_annual
-        for v in M.inputSplitAnnualVintages[r, p, i, t, op]
-    )
-    return indices
-
-
-def LimitTechOutputSplitConstraintIndices(M: 'TemoaModel'):
-    indices = set(
-        (r, p, s, d, t, v, o, op)
-        for r, p, t, o, op in M.outputSplitVintages
-        if t not in M.tech_annual
-        for v in M.outputSplitVintages[r, p, t, o, op]
-        for s in M.TimeSeason[p]
-        for d in M.time_of_day
-    )
-    ann_indices = set(
-        (r, p, t, o, op) for r, p, t, o, op in M.outputSplitVintages if t in M.tech_annual
-    )
-    if len(ann_indices) > 0:
-        msg = (
-            'Warning: Annual technologies included in LimitTechOutputSplit table. '
-            'Use LimitTechOutputSplitAnnual table instead or these constraints will be ignored: {}'
-        )
-        logger.warning(msg.format(ann_indices))
-
-    return indices
-
-
-def LimitTechOutputSplitAnnualConstraintIndices(M: 'TemoaModel'):
-    indices = set(
-        (r, p, t, v, o, op)
-        for r, p, t, o, op in M.outputSplitAnnualVintages
-        if t in M.tech_annual
-        for v in M.outputSplitAnnualVintages[r, p, t, o, op]
-    )
-    return indices
-
-
-def LimitTechOutputSplitAverageConstraintIndices(M: 'TemoaModel'):
-    indices = set(
-        (r, p, t, v, o, op)
-        for r, p, t, o, op in M.outputSplitAnnualVintages
-        if t not in M.tech_annual
-        for v in M.outputSplitAnnualVintages[r, p, t, o, op]
-    )
-    return indices
-
-
-def LimitGrowthCapacityIndices(M: 'TemoaModel'):
-    indices = set(
-        (r, p, t, op)
-        for r, t, op in M.LimitGrowthCapacity.sparse_iterkeys()
-        for p in M.time_optimize
-    )
-    return indices
-
-
-def LimitDegrowthCapacityIndices(M: 'TemoaModel'):
-    indices = set(
-        (r, p, t, op)
-        for r, t, op in M.LimitDegrowthCapacity.sparse_iterkeys()
-        for p in M.time_optimize
-    )
-    return indices
-
-
-def LimitGrowthNewCapacityIndices(M: 'TemoaModel'):
-    indices = set(
-        (r, p, t, op)
-        for r, t, op in M.LimitGrowthNewCapacity.sparse_iterkeys()
-        for p in M.time_optimize
-    )
-    return indices
-
-
-def LimitDegrowthNewCapacityIndices(M: 'TemoaModel'):
-    indices = set(
-        (r, p, t, op)
-        for r, t, op in M.LimitDegrowthNewCapacity.sparse_iterkeys()
-        for p in M.time_optimize
-    )
-    return indices
-
-
-def LimitGrowthNewCapacityDeltaIndices(M: 'TemoaModel'):
-    indices = set(
-        (r, p, t, op)
-        for r, t, op in M.LimitGrowthNewCapacityDelta.sparse_iterkeys()
-        for p in M.time_optimize
-    )
-    return indices
-
-
-def LimitDegrowthNewCapacityDeltaIndices(M: 'TemoaModel'):
-    indices = set(
-        (r, p, t, op)
-        for r, t, op in M.LimitDegrowthNewCapacityDelta.sparse_iterkeys()
-        for p in M.time_optimize
-    )
-    return indices
-
-
-def loop_period_next_timeslice(M: 'TemoaModel', p, s, d) -> tuple[str, str]:
-    # Final time slice of final season (end of period)
-    # Loop state back to initial state of first season
-    # Loop the period
-    if s == M.TimeSeason[p].last() and d == M.time_of_day.last():
-        s_next = M.TimeSeason[p].first()
-        d_next = M.time_of_day.first()
-
-    # Last time slice of any season that is NOT the last season
-    # Carry state to initial state of next season
-    # Carry state between seasons
-    elif d == M.time_of_day.last():
-        s_next = M.TimeSeason[p].next(s)
-        d_next = M.time_of_day.first()
-
-    # Any other time slice
-    # Carry state to next time slice in the same season
-    # Continuing through this season
-    else:
-        s_next = s
-        d_next = M.time_of_day.next(d)
-
-    return s_next, d_next
-
-
-def loop_season_next_timeslice(M: 'TemoaModel', p, s, d) -> tuple[str, str]:
-    # We loop each season so never carrying state between seasons
-    s_next = s
-
-    # Final time slice of any season
-    # Loop state back to initial state of same season
-    # Loop each season
-    if d == M.time_of_day.last():
-        d_next = M.time_of_day.first()
-
-    # Any other time slice
-    # Carry state to next time slice in the same season
-    # Continuing through this season
-    else:
-        d_next = M.time_of_day.next(d)
-
-    return s_next, d_next
-
-
-def gather_group_regions(M: 'TemoaModel', region: str) -> Iterable[str]:
-    if region == 'global':
-        regions = M.regions
-    elif '+' in region:
-        regions = region.split('+')
-    else:
-        regions = (region,)
-    return regions
-
-
-def gather_group_techs(M: 'TemoaModel', t_or_g: str) -> Iterable[str]:
-    if t_or_g in M.tech_group_names:
-        techs = M.tech_group_members[t_or_g]
-    elif '+' in t_or_g:
-        techs = t_or_g.split('+')
-    else:
-        techs = (t_or_g,)
-    return techs
-
-
-def get_loan_life(M: 'TemoaModel', r, t, v):
-    return M.LifetimeProcess[r, t, v]
 
 
 def copy_from(other_set):
